Refactoring to avoid collisions with std lib
There is a standard library package called collecitons. It
is unlikely to cause problems, but I think that having
the subpackage name clash with that could cause issues for
import of collections from inside peytol.

This will require a slight change to phylesystem-api
