+++ extras/tree-list-to-collection.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
-from peyotl.collections import get_empty_collection
-from peyotl.collections.validation import validate_collection
+from peyotl.collections_store import get_empty_collection
+from peyotl.collections_store.validation import validate_collection
 from peyotl import write_as_json
 import sys
