+++ peyotl/api/collections_api.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
-from peyotl.collections.collections_umbrella import TreeCollectionStore, TreeCollectionStoreProxy
+from peyotl.collections_store.collections_umbrella import TreeCollectionStore, TreeCollectionStoreProxy
 from peyotl.api.wrapper import _WSWrapper, APIWrapper
-from peyotl.collections import COLLECTION_ID_PATTERN
+from peyotl.collections_store import COLLECTION_ID_PATTERN
 from peyotl.utility import get_logger
 import anyjson
 import os
